% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detrend_cov}
\alias{detrend_cov}
\title{Detrended Covariance
Functional that returns the detrended covariance between two vectors}
\usage{
detrend_cov(x, y, m)
}
\arguments{
\item{x}{a real valued column vector}

\item{y}{is a real valued column vector}

\item{m}{is the detrending order}
}
\description{
Detrended Covariance
Functional that returns the detrended covariance between two vectors
}
