% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfa.plot.R
\name{dfa.plot}
\alias{dfa.plot}
\title{Detrended Fluctuation Plot}
\usage{
dfa.plot(x)
}
\arguments{
\item{x}{is an object returned from the \code{dfa} function of this package.
Plot parameters are chosen automatically,}
}
\description{
Plot method for monofractal detrended fluctuation analysis
}
