% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dfa}
\alias{dfa}
\title{dfa.noise.out <- dfa(
    x = noise, 
    order = 1, 
    verbose = 1, 
    scales = scales,
    scale_ratio = 2)}
\usage{
dfa(x, order, verbose, scales, scale_ratio = 2)
}
\description{
pink.noise <- fgn_sim(n = 5000, H = 0.9)
}
\details{
dfa.pink.out <- dfa(
    x = pink.noise, 
    order = 1, 
    verbose = 1, 
    scales = scales, 
    scale_ratio = 2)

anticorr.noise <- fgn_sim(n = 5000, H = 0.25)

dfa.anticorr.out <- dfa(
    x = anticorr.noise, 
    order = 1, 
    verbose = 1, 
    scales = scales, 
    scale_ratio = 2)
}
