% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlcca}
\alias{dlcca}
\title{Multiscale Lagged Regression Anlaysis
Fast function for computing MRA on long time series}
\usage{
dlcca(x, y, order, scales, lags, direction)
}
\arguments{
\item{x}{is a real valued vector of time series data}

\item{y}{is a real valued vector of time series data}

\item{order}{is an integer indicating the polynomial order used for 
detrending the local windows}

\item{scales}{integer vector of scales over which to compute correlation. 
Performance is best when scales are evenly spaced in log units. Choosing a
logarithm base between 1 and 2 may also improve performance of regression.}

\item{lags}{integer indicating the maximum number of lags to include}

\item{direction}{string indicating a positive ('p') or negative ('n') lag}
}
\value{
The object returned from the dlcca() function is a list containing rho coefficients for each lag at each of the scales.
}
\description{
Multiscale Lagged Regression Anlaysis
Fast function for computing MRA on long time series
}
