% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fftw_fft}
\alias{fftw_fft}
\title{Fastest Fourier Transform in the west
A simple 1D implementation of the FFTW and its inverse}
\usage{
fftw_fft(in_vec, fft_type)
}
\arguments{
\item{in_vec}{a real or complex valued vector.}

\item{type}{is a character with two values, 'f' = forward, 'i' = inverse}
}
\value{
returns a complex valued 1D FFT
}
\description{
Fastest Fourier Transform in the west
A simple 1D implementation of the FFTW and its inverse
}
