% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgn_sim.R
\name{fgn_sim}
\alias{fgn_sim}
\title{Simulate fractional Gaussian Noise.}
\usage{
fgn_sim(n = 1000, H = 0.7)
}
\arguments{
\item{n}{integer indicating length of desired series}

\item{H}{Hurst exponent ranges between 0 and 1}
}
\value{
A numeric vector of length n.
}
\description{
Simulate fractional Gaussian Noise.
}
