% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fractaldata}
\alias{fractaldata}
\title{A whitenoise, monofractal, and multifractal timeseries}
\format{
An object of class \code{data.frame} with 8000 rows and 4 columns.
}
\source{
\url{https://www.ntnu.edu/inb/geri/software}
}
\usage{
data(fractaldata)
}
\description{
These data include three simulated data to be used for understanding the differences between the various univariate
methods in the dataset to compare whitenoise, monofractal, and multifractal data.
}
\references{
Ihlen, E. A. F. (2012). Introduction to Multifractal Detrended Fluctuation Analysis in Matlab. Frontiers in Physiology, 3. https://doi.org/10.3389/fphys.2012.00141
}
\keyword{datasets}
