% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaafft.R
\name{iaafft}
\alias{iaafft}
\title{Iterated Amplitude Adjusted Fourier Transform}
\usage{
iaafft(signal, N = 1)
}
\arguments{
\item{signal}{is a real valued time serires}

\item{N}{is the number of desired surrogates. Default is 1}
}
\description{
Iterated Amplitude Adjusted Fourier Transform
}
