% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_c}
\alias{lm_c}
\title{Simplef bivariate regression written in c++}
\usage{
lm_c(xs, yr)
}
\arguments{
\item{xs}{a real valued column vector}

\item{yr}{is a real valued column vector}
}
\description{
Simplef bivariate regression written in c++
}
