% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logscale.R
\name{logscale}
\alias{logscale}
\title{logscale
Create logarithmically spaced scales}
\usage{
logscale(scale_min, scale_max, scale_ratio)
}
\arguments{
\item{scale_min}{an integer indicating the minimum scale to be resovled}

\item{scale_max}{an integer indicating the maximum scale to be resolved}

\item{scale_ratio}{a double indiciting the ratio by which scale successive 
scales. For example, scale_ratio = 2 would create a scales increasing by 
a power of 2.}
}
\description{
logscale
Create logarithmically spaced scales
}
