% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdfa.plot.R
\name{mfdfa.plot}
\alias{mfdfa.plot}
\title{Multifractal Spectrum Plot}
\usage{
mfdfa.plot(mf)
}
\arguments{
\item{mf}{an object containing elements related to the mutlifractal 
spectrum derived from Multifractal Detrended Fluctuation Analysis}
}
\description{
Method for plotting various forms of the multifractal spectrum
}
\references{
Kantelhardt et al. (2002). Multifractal detrended fluctuation
analys of nonstationary time series. Physica A: Statistical Mechanics and 
its Applications, 87
}
\author{
Aaron D. Likens (2022)
}
