% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mfdfa_cj}
\alias{mfdfa_cj}
\title{Multifractal Analysis Chhabra-Jensen Method}
\usage{
mfdfa_cj(Timeseries, qValues, scales)
}
\arguments{
\item{Timeseries}{is a real valued time series}

\item{qValues}{real valued vector of q-orders}

\item{scales}{unsigned integer vector of scales to be resolved}
}
\description{
Fast function for computing multifractal analysis using a lesser-known method
for estimating the family of long-range temporal correlations or scaling 
exponents in time series data. This is also a form of multifractal 
analysis that indicates the degree of interaction across temporal scales.
}
