% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poly_residuals}
\alias{poly_residuals}
\title{Polynomial Residuals
Function that fits a polynomial and returns the residuals}
\usage{
poly_residuals(yr, m)
}
\arguments{
\item{yr}{is a real valued vector}

\item{m}{is the detrending order}
}
\description{
Polynomial Residuals
Function that fits a polynomial and returns the residuals
}
