% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_int}
\alias{seq_int}
\title{Integer Sequence
Function that produces a sequence of integers from 1 to N}
\usage{
seq_int(length)
}
\arguments{
\item{length}{is a positive integer that will produce a sequence from 1:length}
}
\description{
Integer Sequence
Function that produces a sequence of integers from 1 to N
}
